Refactor: extract function for canonicalising modeled methods
We don't currently (I think) have an easy way to check whether two modeled methods are the same. I've shuffled some stuff around, so that you can now compare modeled methods by check that their "key" (i.e. JSON-stringified form) is the same.
